Adjust for greater than 16-bit breakpoints

+++ src/main.c
@@ -766,7 +766,7 @@ main(int argc, char **argv)
 			argv++;
 			debugger_enabled = true;
 			if (argc && argv[0][0] != '-') {
-				uint16_t bpVal = (uint16_t)strtol(argv[0], NULL, 16);
+				uint32_t bpVal = (uint32_t)strtol(argv[0], NULL, 16);
 				struct breakpoint bp = { bpVal & 0xffff, bpVal >> 16 };
 				DEBUGSetBreakPoint(bp);
 				argc--;
